+++ modules/fimo_core/src/module_registry.rs
